Factor out SymbolReference
