+++ include/netlist/NetlistEdge.hpp
@@ -1,31 +1,14 @@
 #pragma once
 
-#include <string>
-
 #include "netlist/DirectedGraph.hpp"
 #include "netlist/DriverBitRange.hpp"
+#include "netlist/SymbolReference.hpp"
 
 #include "slang/ast/SemanticFacts.h"
-#include "slang/text/SourceLocation.h"
 
 namespace slang::netlist {
 
 class NetlistNode;
-
-/// Extracted identity of an AST symbol, decoupled from the live AST.
-struct SymbolReference {
-  std::string name;
-  std::string hierarchicalPath;
-  SourceLocation location;
-
-  SymbolReference() = default;
-  SymbolReference(std::string name, std::string hierarchicalPath,
-                  SourceLocation location)
-      : name(std::move(name)), hierarchicalPath(std::move(hierarchicalPath)),
-        location(location) {}
-
-  auto empty() const -> bool { return name.empty(); }
-};
 
 /// A class representing a dependency between two nodes in the netlist.
 class NetlistEdge : public DirectedEdge<NetlistNode, NetlistEdge> {
